test(runtime): centralize advisor test helpers
Share the runtime advisor resolver helpers across test files so provider-based advisor test setup stays consistent and easier to update.
