+++ internal/client/http.go
@@ -496,10 +496,6 @@ func createSessionBoundTransport(provider *typ.Provider, sessionID typ.SessionID
 				model:        model,
 				proxyURL:     provider.ProxyURL,
 			}
-		case ai.IssuerCodex:
-			return &codexRoundTripper{
-				RoundTripper: baseTransport,
-			}
 		}
 	}
