+++ internal/server/anthropic_test.go
@@ -43,7 +43,7 @@ func TestBetaDecode(t *testing.T) {
 	}
 	var jsonString string
 	if err := json.Unmarshal(rawBody, &jsonString); err != nil {
-		panic(fmt.Sprintf("第一次解码失败: %v", err))
+		panic(fmt.Sprintf("first decoding failed: %v", err))
 	}
 
 	// Test that rawBody (a JSON string) can be deserialized into BetaMessageNewParams
@@ -124,10 +124,9 @@ func TestDecode(t *testing.T) {
 	if string(rawBody) == "\"\"" || len(rawBody) == 0 {
 		t.Skip("anthropic_test.txt fixture is empty")
 	}
-	// 2. 第一次解码：得到 JSON 字符串
 	var jsonString string
 	if err := json.Unmarshal(rawBody, &jsonString); err != nil {
-		panic(fmt.Sprintf("第一次解码失败: %v", err))
+		panic(fmt.Sprintf("first decoding failed: %v", err))
 	}
 
 	// Test that rawBody (a JSON string) can be deserialized into BetaMessageNewParams
